+++ app/system/Bakery/Command/ClearCache.php
@@ -12,6 +12,7 @@ use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputOption;
 use UserFrosting\System\Bakery\BaseCommand;
+use UserFrosting\Sprinkle\Core\Twig\CacheHelper;
 
 /**
  * Debug CLI Tools.
@@ -28,7 +29,7 @@ class ClearCache extends BaseCommand
     protected function configure()
     {
         $this->setName("clear-cache")
-             ->setDescription("Clears the application cache");
+             ->setDescription("Clears the application cache. Includes cache service, Twig and Router cached data");
     }
 
     /**
@@ -37,7 +38,60 @@ class ClearCache extends BaseCommand
     protected function execute(InputInterface $input, OutputInterface $output)
     {
         $this->io->title("Clearing cache");
-        $this->ci->cache->flush();
-        $this->io->success("Done !");
+
+        // Clear normal cache
+        $this->io->writeln("<info> > Clearing Illumintate cache instance</info>", OutputInterface::VERBOSITY_VERBOSE);
+        $this->clearIlluminateCache(); //!TODO Add success check. Note that `->flush` won't return true/false because we're using a global tag
+
+        // Clear Twig cache
+        $this->io->writeln("<info> > Clearing Twig cached data</info>", OutputInterface::VERBOSITY_VERBOSE);
+        if (!$this->clearTwigCache()) {
+            $this->io->error("Failed to clear Twig cached data. Make sure you have write access to the `app/cache/twig` directory.");
+            exit(1);
+        }
+
+        // Clear router cache
+        $this->io->writeln("<info> > Clearing Router cache file</info>", OutputInterface::VERBOSITY_VERBOSE);
+        if (!$this->clearRouterCache()) {
+            $file = $this->ci->config['settings.routerCacheFile'];
+            $this->io->error("Failed to delete Router cache file. Make sure you have write access to the `$file` file.");
+            exit(1);
+        }
+
+        $this->io->success("Cache cleared !");
+    }
+
+    /**
+     * Flush the cached data from the cache service
+     *
+     * @access protected
+     * @return void
+     */
+    protected function clearIlluminateCache()
+    {
+        $this->ci->cache>flush();
+    }
+
+    /**
+     * Clear the Twig cache using the Twig CacheHelper class
+     *
+     * @access protected
+     * @return bool true/false if operation is successfull
+     */
+    protected function clearTwigCache()
+    {
+        $cacheHelper = new CacheHelper($this->ci);
+        return $cacheHelper->clearCache();
+    }
+
+    /**
+     * Clear the Router cache data file
+     *
+     * @access protected
+     * @return bool true/false if operation is successfull
+     */
+    protected function clearRouterCache()
+    {
+        return $this->ci->router->clearCache();
     }
 }