+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Twig extensions now implement `Twig_Extension_GlobalsInterface` as required by https://twig.symfony.com/doc/2.x/advanced.html#id1
 - Display element based on permissions for group list/info pages 
 - Factor the admin user creation out of migrations and into it's own Bakery command (See #778)
+- Bakery `clear-cache` command now clears Twig and router cache (Fix #750)
 
 ## v4.1.11-alpha
 - Updated `composer/installers` dependency.